i dont have enough space here lol. solved not being able to get username when having a transfer they were involved in. fixed request bucks and send bucks. both work when used in the console. simplified createTransferFromDto in TransferServiceImpl. transfer DTO coming in have correct account from and to inside of it. added to client/server user classes to implement getting user with their account id

+++ tenmo-client/src/main/java/com/techelevator/tenmo/model/TransferDto.java
@@ -4,12 +4,12 @@ import java.math.BigDecimal;
 
 public class TransferDto {
 
-        private long userFrom;
-        private long userTo;
+        private int userFrom;
+        private int userTo;
         private BigDecimal amount;
         private int transferType;
 
-        public TransferDto(long fromUserId, long toUserId, BigDecimal amount, int transferType) {
+        public TransferDto(int fromUserId, int toUserId, BigDecimal amount, int transferType) {
             this.userFrom = fromUserId;
             this.userTo = toUserId;
             this.amount = amount;
@@ -20,7 +20,7 @@ public class TransferDto {
             return userFrom;
         }
 
-        public void setUserFrom(long userFrom) {
+        public void setUserFrom(int userFrom) {
             this.userFrom = userFrom;
         }
 
@@ -28,7 +28,7 @@ public class TransferDto {
             return userTo;
         }
 
-        public void setUserTo(long userTo) {
+        public void setUserTo(int userTo) {
             this.userTo = userTo;
         }
